pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sheaf"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
sheaf = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
